pyproject.toml
[project]
name = "agentic-webapp"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "outcastgeek", email = "outcastgeek+gitkraken@gmail.com" }
]
dependencies = [
    "python-fasthtml>=0.2.1",
    "pydantic>=2.8.2",
    "sqlmodel>=0.0.21",
    "lancedb>=0.11.0",
    "langgraph>=0.1.19",
    "langchain-anthropic>=0.1.22",
    "langchain-groq>=0.1.9",
    "langchain-openai>=0.1.20",
    "langchain-community>=0.2.11",
    "prompt-poet>=0.0.40",
    "pyzmq>=26.0.3",
    "uvicorn>=0.30.5",
    "termcolor>=2.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.6",
]

[tool.rye.scripts]
webapp_dev = "rye run uvicorn agentic_webapp.webapp:app --reload-dir ."
webapp = "rye run uvicorn agentic_webapp.webapp:app"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_webapp"]
